refactor: 미사용 dto 제거
